add Makefile and CI support
Add a Makefile so that pEpEngineSequoiaBackend can be build using the
usual pEp style without directly invoking cargo, supporting the usual
targets build, install, uninstall, test, clean, all.
Make the project easy to install and uninstall, with a user-settable
prefix.
The Makefile is extensible with a local.conf file, of which we provide
an example.

Add CI support.

Add a .gitignore file.

Co-authored-by: Heck <heck@pep.foundation>
Co-authored-by: Neal H. Walfield <neal@walfield.org>
Co-authored-by: Luca Saiu <positron@pep.foundation>
Co-authored-by: Devan Carpenter <dvn@pep-project.org>
Co-authored-by: Volker Birk <volker.birk@pep-project.org>